+++ database/upgrade-v0.8.5.sql
@@ -1,0 +1,2 @@
+-- v0.8.5
+-- No database structure changes required for this release.